tests: move test utils to tests folder
