+++ packages/chain-helpers/package.json
@@ -30,7 +30,6 @@
   "bugs": "https://github.com/KILTprotocol/sdk-js/issues",
   "homepage": "https://github.com/KILTprotocol/sdk-js#readme",
   "devDependencies": {
-    "@kiltprotocol/testing": "workspace:*",
     "@polkadot/keyring": "^12.0.0",
     "rimraf": "^3.0.2",
     "typescript": "^4.8.3"
