+++ packages/messaging/package.json
@@ -30,7 +30,6 @@
   "bugs": "https://github.com/KILTprotocol/sdk-js/issues",
   "homepage": "https://github.com/KILTprotocol/sdk-js#readme",
   "devDependencies": {
-    "@kiltprotocol/testing": "workspace:*",
     "rimraf": "^3.0.2",
     "typescript": "^4.8.3"
   },
