--- packages/testing/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@kiltprotocol/testing",
-  "private": true,
-  "version": "0.33.2-5",
-  "description": "",
-  "main": "./lib/cjs/index.js",
-  "module": "./lib/esm/index.js",
-  "types": "./lib/cjs/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./lib/esm/index.js",
-      "require": "./lib/cjs/index.js"
-    }
-  },
-  "files": [
-    "lib/**/*"
-  ],
-  "scripts": {
-    "clean": "rimraf ./lib",
-    "build": "yarn clean && yarn build:ts",
-    "build:ts": "yarn build:cjs && yarn build:esm",
-    "build:cjs": "tsc --declaration -p tsconfig.build.json && echo '{\"type\":\"commonjs\"}' > ./lib/cjs/package.json",
-    "build:esm": "tsc --declaration -p tsconfig.esm.json && echo '{\"type\":\"module\"}' > ./lib/esm/package.json"
-  },
-  "repository": "github:kiltprotocol/sdk-js",
-  "engines": {
-    "node": ">=16.0"
-  },
-  "author": "",
-  "license": "BSD-4-Clause",
-  "bugs": "https://github.com/KILTprotocol/sdk-js/issues",
-  "homepage": "https://github.com/KILTprotocol/sdk-js#readme",
-  "dependencies": {
-    "@kiltprotocol/chain-helpers": "workspace:*",
-    "@kiltprotocol/config": "workspace:*",
-    "@kiltprotocol/did": "workspace:*",
-    "@kiltprotocol/sdk-js": "workspace:*",
-    "@kiltprotocol/type-definitions": "workspace:*",
-    "@kiltprotocol/types": "workspace:*",
-    "@kiltprotocol/utils": "workspace:*",
-    "@polkadot/api": "^10.4.0",
-    "@polkadot/keyring": "^12.0.0",
-    "@polkadot/types": "^10.4.0",
-    "@polkadot/util": "^12.0.0",
-    "@polkadot/util-crypto": "^12.0.0",
-    "tweetnacl": "^1.0.3",
-    "uuid": "^9.0.0"
-  },
-  "devDependencies": {
-    "rimraf": "^3.0.2",
-    "typescript": "^4.8.3"
-  }
-}
