+++ packages/vc-export/package.json
@@ -32,7 +32,6 @@
   "devDependencies": {
     "@digitalbazaar/ed25519-signature-2020": "^5.2.0",
     "@digitalbazaar/vc": "^6.0.1",
-    "@kiltprotocol/testing": "workspace:*",
     "rimraf": "^3.0.2",
     "typescript": "^4.8.3"
   },
