+++ tests/testUtils/metadata/spiritnet.ts
